+++ src/super/i-input-text/CHANGELOG.md
@@ -9,6 +9,10 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+#### :rocket: New Feature
+
+* Two new attributes added to `i-input-text`: `inputmode` and `pattern`
+
 ## v3.0.0-rc.199 (2021-06-16)
 
 #### :rocket: New Feature